workers: Make thread-safe
Currently, we document that Workers can only be used from one
reconciler in a process (even though snapshot cloning recently added a
second use of Workers, it obeyed this rule since the node reconciler
is a different process than the cluster reconciler).  But this is
relying heavily on the fact that the default kubebuilder controller
will only process one outstanding resource at a time.  As soon as we
bump MaxConcurrentReconciles, limiting Workers to just one reconciler
is no longer adequate to prevent simultaneous go threads in that
reconciler from both wanting to utilize the map.  Time to add a mutex;
at which point it no longer matters how many reconcilers use workers.

Signed-off-by: Eric Blake <eblake@redhat.com>