crd: Bump to controller-gen 0.16.5
Compared to 0.14.0, this new version has slightly stronger validation
of the resulting generated CRD output (+required annotations are
actually honored!), a more compact RBAC, and a reduction of spurious
comments in the OpenAPI blurb for Status.Conditions.  It also makes it
possible to mark selectable fields, but that will be done in a
separate patch.

Version 0.17.0 is also available, but it requires Go 1.23, and I'm not
ready to make that required yet (Fedora 40 is still on Go 1.22, for
example).

As with the previous patch, although the resulting CRDs are a bit
tighter, I do not view it as a meaningful API change from v0.10.0, but
merely as a preventative measure to avoid breaking assumptions when an
admin manually creates a resource rather than going through CSI calls.

As part of preparing this patch, I also built
quay.io/kubesan/kubesan-tools:0.10.1 with the newer controller-gen
pre-built; that image will also be tagged with :latest once this MR is
accepted.

Signed-off-by: Eric Blake <eblake@redhat.com>