crd: Expose some selectable fields
Kubernetes has a feature gate CustomResourceFieldSelectors, now
enabled by default as of kubernetes 1.31, that allows field selection
of specific fields within a CRD [1].  Expose the fields that make the
most sense for possible filtering (made possible by the previous patch
updating to new-enough controller-gen), such as wanting to list all
resources tied to a specific VG when the system has more than one
Volume Group.  And in my own case, I have pending patches that want to
filter on NBDExports local to a specific node, which gets a lot easier
if kubernetes agrees that NBDExport.Spec.Host field is selectable.

[1] https://kubernetes.io/docs/tasks/extend-kubernetes/custom-resources/custom-resource-definitions/#crd-selectable-fields

Intentionally missing from this patch: I'd really love to be able to
select on Volumes that are block vs filesystem.  But to do that, I
need to make an incompatible API change to make VolumeType a proper
discriminated union, so it will have to wait for v0.11.0.  Likewise,
once we start allowing snapshots to be created from a filesystem, they
will need to record whether they are a block or filesystem (a new
volume must match the type of the snapshot it is created from, because
we absolutely do not want to allow the user to create a filesystem
volume from a block snapshot).

Signed-off-by: Eric Blake <eblake@redhat.com>