crd: Mark all fields +required or +optional
Make several CRD robustness changes, as suggested by
https://ahmet.im/blog/crd-generation-pitfalls/.

Adding a package level marker to default all fields to required as a
last-ditch backup that will help force us to think about future
additions.  In doing so, reserve the kubebuilder-specific spelling of
+kubebuilder:validation:Required to the package marker, and use the
spelling +required everywhere else, as that is known to a wider set of
tools.  Then make sure every field has either +optional or +required,
in general consistent with whether the field uses omitempty, and
require .spec to exist (since an optional sub-struct with required
members is not validated if the parent struct is omitted).

Any field that should never be empty should explicitly mention a
minLength of 1, even if the regex Pattern also required a non-empty
match to validate.

This change should be benign for v0.10.0 compatibility (it adds more
restrictions, but all existing CRs created by CSI calls already
complied, and making it harder for an admin to manually create a CR
without CSI that violates any assumptions we made is a good thing).
In particular, an admin can no longer create a CR without a Spec
entry.

Signed-off-by: Eric Blake <eblake@redhat.com>