tests: Another improvement to kcli cache
I managed to get the snapshotter code installed before creating the
kcli cache, but didn't realize that the --use-cache setup triggers
another code path that also tries to install the snapshotter.
Thankfully, it's idempotent, but it's wasted effort, and easy enough
to avoid - if we are loading from the snapshot, we no longer need the
snapshotter on any subsequent code path.

Fixes: 3164b774 ("tests: Deploy snapshot CRDs prior to creating cache")
Signed-off-by: Eric Blake <eblake@redhat.com>