nbd: Better reconcile filtering
Set up a predicate to pre-filter things, since the reconciler only has
to act on exports on the current node, and only when the Spec has
changed.

Test by running `tests/run.sh --pause-on-stage tests/t/nbdexport.sh`
and watching the logs; the node(s) unrelated to the server no longer
wake up in the nbdexport reconciler loop, and the node handling the
start or stop of the export no longer has back-to-back reconciles
after updating Status.

(We could possibly add a node predicate on ThinPoolLv's node
reconciler, but we can't use the GenerationChanged predicate on any of
Node, Snapshot, or ThinPoolLv because those actively require waking up
due to status.Conditions changes for coordination between reconcilers,
and/or when a secondary object changes requiring the primary object to
be re-reconciled even though its generation did not change.)

Signed-off-by: Eric Blake <eblake@redhat.com>