nbd: Make q-s-d a sidecar, for cleaner shutdown
When a pod has more than one regular container, there are no
guarantees as to which container will get SIGTERM first from kubelet
unless the pod declaration included some preStop hooks to inject
sleeps to order things.  But a preStop hook is action at a distance;
we can do better.  When a pod is a Kubernetes Sidecar (that is, an
initContainer pod with restartPolicy of Always), kubelet guarantees
that it will not send SIGTERM to sidecar containers until after all
main containers have exited (including on forceful exit if the pod has
reached the expiration of its graceful timeout).  Kubernetes is also
nice in that once a pod is marked for deletion (which is what triggers
kubelet to send SIGTERM to the manager in the first place), a sidecar
CAN be terminated early and won't be restarted.

How does that relate to us?  We absolutely do not want q-s-d
terminating while the manager is still trying to utilize q-s-d to
coordinate RWX volume access - yet before this patch, nothing was
preventing kubelet from sending SIGTERM to q-s-d while the manager was
still doing work.  Promoting q-s-d to a sidecar means that we are now
ensured that even if we do nothing else, it will now outlive the
manager, so we can still rely on q-s-d shutting down gracefully when
kubelet sends it SIGTERM.  Optionally, we can allow q-s-d to be shut
down even sooner, and in fact, we can set up a channel that will be
closed at the first point where it is safe to shut down q-s-d
(ie. when there are no client connections).  Actual use of the clean
shutdown function won't happen until the next patch improves the
manager's SIGTERM handling; this patch is focused just on q-s-d
interactions.

Note that even when this new function is called, the QMP "quit"
command is not always sent - depending on how fast the manager cancels
things after SIGTERM is seen, the main thread may have exited before
the go thread created here can run to completion.  But that's okay -
as soon as main exits, kubelet sends SIGTERM to q-s-d, with the same
effect as a QMP "quit", and safe when we are sure there are no clients.

Signed-off-by: Eric Blake <eblake@redhat.com>