nbd: Delay shutdown until all exports are stopped
Our design needs to be robust to the manager Pod restarting, since
that is the pod running qemu-storage-daemon.  We do not want to be
serving any NBD exports at the time we know the manager pod needs to
quit, because that would break any RWX volumes depending on the
server.  Our Volume code can already handle temporary pauses of client
nodes for migration purposes, so it's time to put that code into
service for also handling temporary pauses in the availability of
q-s-d because we know the Pod is being deleted (presumably to be
replaced by the same or a newer version of the manager Pod, thanks to
the power of kubernetes' DaemonSets).

It is worth noting that with our DaemonSet, the new pod will not start
execution until the old pod has completely terminated.  This is good
in the sense that there are never two controllers running on the same
node competing to make the same status changes, but it does mean that
if we don't quit the old pod quickly, the new pod cannot start until
the old pod's grace period expires, leaving volumes suspended for the
entire time of the handover.

The previous patch set up q-s-d to inform us when it is at a safe
point, now we need to wire up utilizing that function. The Kubebuilder
Manager already accepts a master context that it passes to all
controllers, and provides ctrl.SetupSignalHandlers() to create such a
context - however, that context does not have any way to inject
anything between the receipt of SIGTERM and when it actually calls
cancel().  But it's easy enough to replace that function pointer to
instead use our own implementation that adds the NBD cleanup we need.

The rest of this patch is then setting up what we DO when manager.go
calls our replacement function for setting up the master context.  In
addition to registering a signal handler, we need to call client.List
filtered to the NBDExports on the current node (if any), and then
update those exports to start the sequence that will eventually let
the last one release the channel returned by nbd.StopServer().  Since
this cleanup trigger is not executing in an actual controller, there
won't be any automatic exponential backoff, so it needs to retry on
conflicts to maximize the chance that it actually works.

The best way I found to test this patch: Use `tests/run.sh
--pause-on-stage tests/t/nbd-pod-restart.sh`, then connect a second
terminal to the same $KUBECONFIG at the point where an NBDExport is
running.  In one terminal, `kubectl logs -n kubesan-system -f node-XXX
| less` will collect the logs from a given node controller all the way
through its death, then in the other terminal, `time kubectl delete -n
kubesan-system pod node-XXX` shows how long things took.  The `less'
in the first terminal then lets you spot if all the cleanup messages
occurred in the right timeline (and be sure to observe what happens
both for a node with an NBDExport, and a node without).

Fixes: https://gitlab.com/kubesan/kubesan/-/issues/87
Signed-off-by: Eric Blake <eblake@redhat.com>