csi: Switch from client-go to controller-runtime
client-go may be the built-in kubernetes client, but it is very
low-level.  If you want to build up a shared cache via informers, you
have to duplicate the boilerplate yourself - and yet, we NEED a cache
to check things from a client-side store, because without one, every
call to watch.UntilWithSync blocks the entire REST endpoint until the
watch is satisfied from the server.

Thankfully, controller-runtime already has everything we need for
easily creating a client-side store and running informers under the
hood.  The fact that we're already using it in our manager code is
another benefit - there's no need to have to swap between engines just
because we're swapping between directories of the project.

See also this article[1], which details another project's switch to controller-runtime.

[1] https://medium.com/@timebertt/kubernetes-controllers-at-scale-clients-caches-conflicts-patches-explained-aa0f7a8b4332

I found it to be slightly better performance if we prime the cache by
requesting an object of each type as soon as the process starts,
rather than waiting for the first CSI call.  But since our node
plugins don't ever interact with Snapshot objects, it required a
slight refactor to the client creation code to take a parameter
describing whether the cache will be used for a cluster.

Note that since there are informers under the hood maintaining the
cache, it would still be possible to use edge-based triggering (as
soon as the watch stream sees the change to the Volume we are
watching, have the informer wake up a channel registered by the gRPC
call in addition to updating the cache) - but that is more complicated
to write and reason about.  Instead, the simpler approach is the
kubernetes philosophy that everything eventually stabilizes, so just
periodically polling the cache rather than directly hooking into the
informers will, eventually, see the right level-based results.  The
polling done here is NOT making any http calls out to etcd (it is only
reading from the cache, and other gRPC calls can now make write
requests while the informer's watch request is still ongoing), but we
also know there are exponential backoffs in the manager, so we can
slow down our polling over time to avoid churning on CPU while we
wait.  I picked polling that ranges from 0.25 up to 3 seconds, ramping
less aggressively than doubling, which results in up to 9-10 polls
within the default 15-second window that csi-provisioner expects of us
(depending on jitter); we may still choose to tune these further
later.

In testing, I was able to see better parallel response behaviors (for
example, this was the first time where I was able to observe two
overlapping blkdiscard calls in the manager for two linear block
volumes created at the same time without the unsafe fast wipe policy;
where previously the Volume for the second wasn't acted on until the
Volume for the first had stabilized).  But I also saw csi-sanity take
longer to complete (polling instead of edge triggering adds some
latency for individual operations, even if throughput for parallel
operations improves).

This addition feels like some duplication; there are now four separate
wait.Backoff calls.  The next patch will do some further refactoring
to get that down to one.

Signed-off-by: Eric Blake <eblake@redhat.com>