csi: Consolidate backoff wait loops into one implementation
Rather than open-coding four different wait loops, exploit the fact
that we have a common Object interface and route all waits through the
same code.  This slightly changes the timeout characteristics of
waiting for deletion (a bit more frequent), but as argued in the
previous patch, this is not resulting in more http traffic to etcd.

If we need to further tune things, now we have just one place to
touch.

Signed-off-by: Eric Blake <eblake@redhat.com>