csi: Enable leader election on cluster controller
While every node has a node manager running, we have, up until now,
had only one controller plugin.  If it happens to be running on a node
that fails, we want the cluster to immediately start servicing
cluster-wide CSI reconcilation from another node, rather than dealing
with a potentially large gap in time with no service before the
Deployment notices and fires up a replacement pod on another node.
The best way to do that is to have a replica count larger than one,
coupled with a topology spread constraint that forces the scheduler to
separate replicas onto different nodes (while a one-node cluster is
possible, it is not the target for KubeSAN, as the whole point of a
shared VG is to have storage shared between more than one node).  But
since only one of those replicas should be performing reconciliation
at a time, we also need to enable the built-in leader election
services provided by Kubebuilder.

Note that our gRPC code is idempotent and does NOT need leader
election; put another way, it is not a kubernetes controller trying to
make exclusive-ownership changes to externally visible resources, but
rather is only making .spec changes to internal resources as directed
by CSI calls.  On the other hand, 4 of the 5 CSI sidecars that we use
DO have controller semantics (each is monitoring, modifying and
updating cluster-wide resources based on the results of gRPC calls;
only the liveness-probe container does not need RBAC access to
external resources), and only work if at most one replica is making
the CSI calls relevant to that sidecar's purpose.  What's more
interesting is that these controllers are not required to settle their
leader elections onto the same pod; for example, the provisioner could
be serving CreateVolume on one replica while the snapshotter is
serving CreateSnapshot on another.

In order to enable leader elections, we need additional command line
arguments to each affected container, as well as an additional RBAC
binding (although we can reuse the same RBAC Role that the manager
leader election uses, rather than needing to create a new Role).

One other wrinkle - tests that previously worked by assuming there was
exactly one csi-controller-plugin pod running on the cluster need to
be updated to deal with the new reality of there being more than one
pod, but only one of them active.  Internal function
__ksan-get-pod-name and exported ksan-csi-controller-plugin will only
target the first such pod, without regards to how many leader
elections that pod won (anywhere between 0 and 4). Meanwhile, the
csi-sanity test doesn't care whether the pod that it picks has the
lease, because it is directly hitting the gRPC port (which is always
active, and not subject to leader election) rather than utilizing
anything associated with the CSI sidecars.

You can now see the result of the elections by inspecting `kubectl get
-n kubesan-system lease`.

Signed-off-by: Eric Blake <eblake@redhat.com>