thinpool: Make copy after successful Update
It turns out that client.Update not only writes, but reads back what
the server may have adjusted things to based on mutating webhooks.
Among other things, this means that the metadata volume.Generation
number is different after Update than it was before - and we WANT that
newer generation number to be the basis of any further edits in the
same control path.

Fixes: b20c872b8 ("thinpool: Update oldThinPoolLv on client.Update()", v0.10.0)
Signed-off-by: Eric Blake <eblake@redhat.com>