csi: Enforce ThinLv spec state transition
Fix a TODO by adding CEL validation to ensure that a clueless admin
(or a bug in our code) can't accidentally try to revive a ThinLV once
it has been marked for deletion.

Although this changes a corner case of the API, it should not be
triggered in practice and makes our code safer, so it is fine for
v0.10.x.

Signed-off-by: Eric Blake <eblake@redhat.com>