thinpool: Simplify ThinLv removal state machine
The existing code for removing a ThinLv has a number of back and
forths, each of which requires time for the api writes to be reflected
back to the cached readers via another round of Reconcile():

owner (volume or snapshot):             thinpool:
Reconcile()
  client.Get()
  set spec.thinlv[n].state = Removed
  client.Update()
                                        Reconcile()
                                          client.Get()
                                          call lvremove
                                          set status.thinlv[n].state = removed
                                          client.Status().Update()
Reconcile()
  client.Get()
  if status.thinlv[n].state != removed
    return WatchPending
Reconcile()
  client.Get()
  delete spec.thinlv[n]
  client.Update()
                                        Reconcile()
                                          client.Get()
                                          remove status.thinlv[n]
                                          client.Status().Update()
Reconcile()
  client.Get()
  if status.thinlv[n] still present
    return WatchPending
Reconcile()
  client.Get()

We intentionally want to refuse to delete a ThinLv merely because it
disappeared from the source array (mostly because it makes idempotency
easier, in case we have to re-start a request from Volume or Snapshot;
but also because it offers protections from a clueless admin that
manually tries to update .spec), and only do so when the .spec has set
state to Removed.  But we can skip setting .status state to Removed,
and instead just shrink the .status array.  This gives us less waiting
on another round of reconcile, and can lead to faster deletion
performance.

For now, the CRD keeps the enum value Removed for status, even though
it no longer sets that value - this is because for v0.10.1, we want to
be backwards-compatible to an existing ThinPoolLv that might already
have that status in place.  But for v0.11.0, I want to rework the CRD
to be easier to use (such as dropping Status.Active and making Status
itself the enum rather than Status.Name), where an API break is okay
because we are still in alpha.  Also, the comparison of state names
can be made a private function, since it only had one callsite.

Signed-off-by: Eric Blake <eblake@redhat.com>