thinpool: Cope with thinlv deletion during pool deactivation
The ThinPoolLv node controller was not idempotent, which showed up as
intermittent races causing hangs in csi-sanity.  If we managed to
remove a ThinLv (such as when deleting a Snapshot while a Volume still
remained in the pool) but the corresponding Status.ThinLv[n].State has
not yet been pruned, yet the other ThinLv was inactive, then on entry
to thin pool reconciliation we were trying to call 'lvchange -an
deletedlv'; and since that fails when the lv is not found, no progress
would be made.

Fix it by combining two iteration loops into one, and adding special
cases for removed LVs - if one is already known to be removed, we do
not want to revert its status back to Inactive; meanwhile, if we have
not yet updated status to explicitly mark something removed but we are
deactivating the pool, we know that a missing LV will not prevent
deactivation (later code will still flag the case where an admin
force-removed an LV behind KubeSAN's back).

Fixes: 8ba61c17 ("manager: Add thin LV removal case to thinPoolLvNeedsActivation()", v0.10.0)
Signed-off-by: Eric Blake <eblake@redhat.com>