linear: Use shared activation for zeroing
For linear volumes, we do the zeroing on creation or extension in the
context of the cluster controller.  But what happens if the cluster
controller changes nodes, such as during a rolling update to a new
version of KubeSAN, while zeroing was underway?  The blkdiscard on the
old pod is killed, but the lvm is left stranded in an exclusive
activated state, with no record of which node had activated it, so the
cluster controller on the new pod is unable to take over shared
activation, and there is no breadcrumbs left on which old node needs
to deactivate.

Better is to use shared activation for image zeroing.  Even if we
strand a blkdiscard process and leave the LV activated on an old node,
the moment that the volume moves to Conditions["Available"] means that
the node controller can then deactivate to be in line with
spec.AttachToNodes.

Signed-off-by: Eric Blake <eblake@redhat.com>