commands: Add optional binding tag to lvcreate
An upcoming patch wants to plumb up the ability to document a
human-readable name that is being bound to an LV (either the volume id
from csi-sanity, or the human-supplied PVC name that got turned into
the opaque "pvc-UUID" by the PV of kubernetes).  To prepare for that,
update the signature of LvmLvCreateIdempotent to take an arbitrary
string, which if non-empty, will become an LV tag (use lvs
--devicesfile=vgname -o +lv_tags to see it); complete with munging of
characters that cannot appear in an LV tag.  However, for now, all
callers pass an empty string.

Signed-off-by: Eric Blake <eblake@redhat.com>