crd: Add optional Binding field for recording original name
Add in some plumbing to set an optional tag on an LV during creation.
Nothing sets binding yet (that will be the next patch), but this sets
the stage to get the information from CSI down to the LV creation.

This is a change to the API, but since the new field is optional, it
is safe to do in the v0.10.x branch (older volumes simply lack the
binding information).

Signed-off-by: Eric Blake <eblake@redhat.com>