manager: Allow more memory to cluster manager
In stress tests on an OpenShift cluster, creating 20 linear volumes in
short succession triggered OOMKilled failures of the cluster manager
as it started spawning blkdiscard child processes; but running with a
larger limit and watching the metrics page, the container seemed to
top out at about 135Mi in memory.  Since that is just above the 128Mi
threshold, it explains the OOM death.  Bump the limit by 4x to allow
some more breathing room.

It is entirely likely that better tuning numbers will be chosen later
with better observation of performance under real usage scenarios; but
this patch at least avoids the worst congestion seen in early testing.

Signed-off-by: Eric Blake <eblake@redhat.com>