volume: Reduce errors reported when deleting thin volume
While inspecting logs, I noticed that when a Volume was deleted by
CSI, there was a race where node controllers could complain about
"unexpected missing blob".  Per the life-cycle, we expect the
ThinPoolLv to exist until we remove the ThinLV in the volume cluster
reconcile, and that in turn waits until the volume is not staged on
any nodes.  But once that happens, a node reconcile can still wake up
to a Volume that has been marked for deletion and where the ThinPoolLv
is gone or at least dropped the ThinLV, and that is not an error.

Even if an admin tries to delete a Volume prior to CSI being done with
things, the volume finalizer won't be removed until the ThinLv is
successfully unmounted (ie. the removal will still be deferred until
after the volume is unstaged).  However, the reconciler for linear
volumes already refuses to activate on a node once the volume is
marked for deletion (even when the marking is from an overactive admin
rather than from CSI), so we can do likewise for thin volumes.

Signed-off-by: Eric Blake <eblake@redhat.com>