crd: Drop useless patchStrategy markers
Kubernetes does not support strategic merges on CRDs, so any mention
of patchStrategy or patchMergeKey has no impact to the generated
OpenAPI yaml and only serves to mislead.  It does, however, support
server-side apply (even if we aren't yet using it), and that honors
+listType and +listMapKey, so leave those annotations in place.

As proof that the old markers didn't work, note that in ThinPoolLv,
two of the three uses supplied a bogus patchMergeKey=type (when there
is no 'type' member of the object in the list, and where
listMapKey=name was correct).

Fixes: https://gitlab.com/kubesan/kubesan/-/issues/93
Signed-off-by: Eric Blake <eblake@redhat.com>