crd: Apply common labels to dynamically created resources
All of the resources in our deployment are labeled with the k8s common
labels app.kubernetes.io/{name,version}.  It is also desirable to mark
the resources we create at runtime with the same information, as it
can be useful to see everything that kubesan has touched, and which
version of kubesan did the creation (especially when attempting to
upgrade a cluster in place).

The release script needs a minor tweak to cope with the additional
whitespace that `go fmt` insists on in config.go.

Fixes: https://gitlab.com/kubesan/kubesan/-/issues/78
Signed-off-by: Eric Blake <eblake@redhat.com>