deploy: Munge short names in devel deployment
The devel deployment intentionally creates CRDs with the same Kind but
a different group name.  On any cluster with parallel KubeSAN
installed, this means that `kubectl get -A volumes` is now ambiguous
on whether it will show results only for volumes.kubesan.gitlab.io or
for volumes.devel.kubesan.gitlab.io.  Having to type things out to
disambiguate is doable, but annoying.  But we can use short names to
our advantage - as long as the two CRDs do not share a common
short-name, it is possible to use `kubectl get -A vol` as the short
name for normal installation, and `kubectl get -A d-vol` as the short
name for devel installation, which ultimately is less typing.

(I tried to see if kustomization replacements would allow an empty
string delimiter, but that replaces the entire string.  So this
instead adds a prefix of 'd' with an injected delimiter of '-'.)

Signed-off-by: Eric Blake <eblake@redhat.com>