linear: Move blkdiscard work into node reconciler
As hinted at in earlier patches, it is time to move the blkdiscard
work into the node reconciler.  This means that CreateBlob() can
finish much earlier, and we now are one step closer to also enabling
creation of linear volumes by copying contents.

I noticed one potential race condition: when a Volume is marked for
deletion, the cluster controller may attempt to remove the blob prior
to when the node controller has had a chance to cancel outstanding
blkdiscard work.  However, this is not a fatal problem: lvm will error
out about the device still being in use but exponential backoff will
eventually retry at a point in time where the blkdiscard has been
canceled.

I also changed the creation to utilize lvm's ability to auto-zero the
first 4k (note: 4k, not 4M that we were previously doing), which
requires doing the lvcreate with activation.

There are still some cleanups that will be done in the next patch,
such as removing the now-unused commands.WithLvmLvActivated, as well
as removing workers from the cluster manager.

Signed-off-by: Eric Blake <eblake@redhat.com>