tests: Refactor mode support to allow linear-from-thin
Upcoming patches want to enable creation of a linear volume cloned
from a thin source (snapshot or volume).  To do that, it will be a lot
easier if the testsuite always has two StorageClasses installed, one
for each mode, and where the helper functions ksan-create-*volume can
now allow a temporary override to mode compared to what the rest of
the test is using.  In particular, the clone test needs some tweaks to
ensure that the clone source is always thin, even when the destination
follows $mode.

Signed-off-by: Eric Blake <eblake@redhat.com>