tests: Let csi-sanity pass under --mode Linear
We can't create a snapshot of a linear volume, so any csi-sanity test
that depends on a snapshot existing will fail.  But since we know that
limitation, it's better to skip that test than to fail csi-sanity.
The failures only started happening when we started advertising
snapshot support, but since our CI is not testing --mode Linear, we
didn't notice until now.  But as of this patch, the release process
also ensures that we don't regress on Linear functionality.

Fixes: baec573d ("tests: Noew that csi-sanity passes, hard-fail on regressions")
Signed-off-by: Eric Blake <eblake@redhat.com>