csi: Prevent clone of deleted source
If a volume or snapshot is currently marked for deletion, treat it the
same as missing from the perspective of CSI trying to copy from that
source.  Similarly, an attempt to idempotently re-create something
marked for deletion should fail (it already did for volumes in
validateVolume(), but not in snapshots).

Signed-off-by: Eric Blake <eblake@redhat.com>