clone: Fix cleanup when source is deleted first
Because a clone operation is long-running, the source Volume or
Snapshot of a clone can be deleted in between the time that we first
create the temporary snapshot and when we finally finish the clone and
are ready to clean up.  Kubernetes tries to prevent this (it is able
to warn when a second pvc that is not yet bound still depends on the
first), but can't prevent a determined user from using `kubectl delete
--force`.  In general, we anticipated this scenario - after all,
that's why we created a temporary clone, so that we aren't delaying
the deletion of the original.  But we had two slight flaws:

First, getDataSourceBlob depends on the Volume or Snapshot in order to
determine the appropriate name needed to prepare the ThinBlobManager
we use to delete the clone.  If the original source resource is
deleted first, then the later getDataSourceBlob returns nil, which in
turn skips the deletion of the ThinLv, stranding both the source
ThinPoolLv and the destination Volume in permanent limbo even when
they are marked for deletion.

Second, note that before cloning, we were relying on k8s garbage
collection to automatically mark the ThinPoolLv for deletion when the
last OwnerReference was unlinked at the time the owner object was
deleted.  But now that the clone owner object is voluntarily giving up
its owner reference at the point the clone completes (rather than
because the target was marked for deletion), we can end up with a
ThinPoolLv that has zero ThinLvs but is never marked for deletion.

The fix for the latter is easy: add code to explicitly mark a
ThinPoolLv for deletion if we have just removed its last ThinLv.  For
the former, we need somewhere stable that we can reference that will
outlive the ThinLV itself; the perfect candidate for that is a label
on the destination Volume (I also tried using lvs on the ThinLV itself
to look up the pool - but that has the lifecycle problem of no longer
being able to look up the pool once the ThinLV is deleted).  We
already have a label, although right now it is pointing to a string
that we can always regenerate without having to look up the label, and
our cleanup of that label needs to be delayed a bit later.  And once
we move where the label is removed, it is also easier to move its
creation to match for symmetry.

Our original thoughts for the clone design was to have multiple clone
destinations all share the same label back to a common source to make
it possible to tell when the common source was no longer being cloned,
but that was before Stefan figured out that since we already have a
thin pool, it is dirt cheap to always take an lvm snapshot of the
source (whether the source is a Volume or Snapshot), at which point
every source is unique, and so there is no benefit to doing
client.List because there will be no other destination Volume sharing
our source.  It's nice to know that the label idea still works, just
for a slightly different purpose.

Fixes: a5647ff ("manager: Implement cloning from snapshots and volumes", v0.10.0)
Fixes: e6698199 ("manager: Add cloning API to BlobManager", v0.10.0)
Signed-off-by: Eric Blake <eblake@redhat.com>