thinpool: Be more precise during node reconciliation
The old code for node reconciliation is very fragile: it depended on
no one ever setting Spec.ActiveOnNode different from a non-empty
Status.ActiveOnNode while any of the ThinLvs are requested to be
active (if that happens, the old node refuses to deactivate, but the
new node fails to activate, leaving the pool and all its constituent
thinlvs wedged).  And it was dependent on Status keeping track of
which node last activated the pool.  thinpoollv.go had to work around
this fragility by ensuring that even when other parts of the code want
to change which node is active, the ThinPoolLv.Spec itself forcefully
deactivates all ThinLVs and waits for Status.ActiveOnNode to change to
"" before permitting the rest of the code to reactivate.

The old code also returned a bool for whether the pool should be
deactivated in a defer block; but this was mostly useless - either we
just activated because of an active thinlv (and don't call defer), or
we just deactivated (generally no need to do it again[*]), or we
aren't the node listed in spec (we want the deactivation, but if it is
not a no-op, we have no way to report if it failed because a thinlv
was still active).  In general, changing the state of a system that
should normally be visible through Status should not be done in a
defer. [*] The defer did have one benefit of working around an lvm bug
where an lvremove of a thinlv could accidentally leave the thin pool
locked but not activated; but instead of using a defer, we can keep
that workaround by an explicit invocation.

Better is to ask LVM whether a given node has the pool active.  Using
`lvs -S lv_active=1` would get this information, but requires a
round-trip through lvmlockd obtaining the locks.  More efficient is to
recognize that every activated lvm device has a corresponding local
device-mapper device; if `dmsetup info -S` produces output, then the
pool is active.  From there, recognize that on activation, we may
first need to wait for another node to deactivate; and on
deactivation, it is more useful to return whether the current node
should remain in charge of the rest of the reconcile, instead of
whether there were any thinlvs marked active.

Technically, lvm will lock the pool for any thinlv activity that
requires that lock (such as lvcreate, lvchange -ay of a thinlv, etc);
so we never need to explicitly activate the pool; in fact, dmsetup
info lists up to two separate rows for the thin pool: vg-pool if the
pool is explicitly activated, and vg-pool-tpool if the pool is locked.
But it is certainly easier to debug if we activate the pool as a
bookend around any time any of its thinlvs are activated.

With this patch in place, the ThinPoolLv reconciler is now more
level-based, rather than requiring a precise state machine of
edge-based events.  The next couple of patches can simplify the work
that thinpoollv.go has to do as a result.

Signed-off-by: Eric Blake <eblake@redhat.com>