volume: Grab better size when attaching NBD
Trying to size a device-mapper larger than the underlying device is
not a good idea.  Although the code is designed to restart a larger
NBD server any time the underlying LV is expanded, there is a
possibility that we can hit a race window where the existing NBD
server is still serving a smaller size than what we read from the
ThinPoolLv.  In that case, we should resume the linear wrapper with
the smaller size.

Fixes: acc7b6ce ("volume: Query actual lvm size", v0.10.0)
Signed-off-by: Eric Blake <eblake@redhat.com>