volume: Prevent NBD migration while cloning is active
Fix a TODO comment in the volume code: if there is a clone going on,
the background dd task is using direct access to the LV and not set up
to go through the device-mapper, therefore it is unable to utilize an
NBD connection.  Changing the active node of a ThinPoolLv in such a
scenario will not work.  But since cloning is uncommon, what we CAN do
is arrange for the volume code to set up an NBD server from the node
where the pool is active (whether or not that node is staged), by
ensuring that the set of nodes interested in a thin pool's activation
always favors the pool's current node if the pool is use by a clone.
The NBD server will gracefully migrate to the staged node once cloning
is no longer pinning the preference.

In short, Spec.AttachToVolumes is the list of nodes that determines
whether ThinLvSpec should be active, but thinpoollv.InterestedNodes()
is the list of nodes that determines what thinPoolLv.Spec.ActiveOnNode
should be.  The latter is either the former (possibly in a different
order), or a superset of the former.

Furthermore, reconcileThinDetaching should not wait for the ThinPoolLv
to deactivate (it may still be active on the node because of a pending
clone), but merely for the ThinLV to go offline (and even then, only
if this was the last node staging the ThinLV).  This is almost the
same code as thinpoollv.go DeactivateThinLv, but we have to update the
ThinPoolLv for the change in ActiveOnNode even if we aren't
deactivating the LV.

I was able to test this by introducing an artificial sleep into the dd
worker task to give myself even longer.  Then, with a source volume in
place, I requested creation of a second volume that clones the first,
observed what node that work starts on (ie. the node where the cluster
manager is running), and then intentionally created a pod pinned to a
different node and using the first pvc as a mount.  Even though only
one node is staged, this patch is enough to force the node where the
dd is still happening to spin up an NBD server and allow the pod to
connect through an NBD client, whereas pre-patch, the act of trying to
create the pod caused mayhem when the thinpoollv.go code tried to
change the clone volume to inactive while it was in use.  But I'm not
sure how easy it would be to automate that into the testsuite.

Fixes: 65934552 ("nbd: Add NBD cleanup support to Volume", v0.9.0)
Signed-off-by: Eric Blake <eblake@redhat.com>